rosbag_processor/requirements.txt
asttokens==2.4.0
backcall==0.2.0
bagpy==0.5
bitarray==2.8.2
bitstring==4.1.2
catkin-pkg==1.0.0
contourpy==1.1.1
cycler==0.12.1
decorator==5.1.1
distro==1.8.0
docutils==0.20.1
exceptiongroup==1.1.3
executing==2.0.0
fonttools==4.43.1
gnupg==2.3.1
importlib-resources==6.1.0
iniconfig==2.0.0
ipython==8.12.3
jedi==0.19.1
Jinja2==3.0.3
kiwisolver==1.4.5
lz4==4.3.2
MarkupSafe==2.1.3
matplotlib==3.7.3
matplotlib-inline==0.1.6
numpy==1.24.4
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
pluggy==1.3.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py3rosmsgs==1.18.2
pycryptodomex==3.19.0
Pygments==2.16.1
pyparsing==3.1.1
pyserial==3.5
pytest==7.4.2
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
rosbags==0.9.16
rosbags-image==0.9.1
rospkg==1.5.0
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.8
seaborn==0.13.0
setuptools-scm==8.0.4
six==1.16.0
stack-data==0.6.3
tomli==2.0.1
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
wcwidth==0.2.8
zipp==3.17.0
zstandard==0.21.0
